requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
plotly
Pillow
statsmodels
SQLAlchemy
